refactor: 중복된 쿼리 삭제
