+++ src/entities/submitCode/index.ts
@@ -1,3 +1,9 @@
+/**채점 */
+export { useSubmissionForResultMutation } from './submission/model/mutation/submitCode.mutation';
+export type { ISourceCode } from './submission/model/mutation/submitCode.mutation.type';
+export type { ISubmitCodeRequest } from './submission/model/mutation/submitCode.mutation.type';
+
 /**코드리뷰 */
-export { useGetTokenCountQuery } from './submission/model/query/getTokenCount.query';
 export { useISubmissionForReviewMutation } from './submission/model/mutation/submitCode.mutation';
+export type { ISubmissionReviewRequest } from './submission/model/mutation/submitCode.mutation.type';
+export type { ISubmissionReviewResponse } from './submission/model/mutation/submitCode.mutation.type';
