--- src/entities/submitCode/submission/model/query/getTokenCount.query.types.ts
@@ -1,3 +1,0 @@
-export interface IReviewToken {
-  reviewToken: number;
-}
